backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
sqlalchemy>=2.0
pydantic[email]>=2.7
python-dotenv>=1.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]>=3.3.0
psycopg2-binary>=2.9.9
pydantic-settings>=2.1.0
python-multipart>=0.0.6
pdfkit>=1.0.0
jinja2>=3.1.2
python-magic>=0.4.27
aiofiles>=23.2.1
bcrypt==3.2.2
cryptography>=41.0.0
email-validator>=2.0.0
python-dateutil>=2.8.2
pytz>=2023.3
aioredis>=2.0.1
prometheus-client>=0.17.1
sentry-sdk[fastapi]>=1.29.2
